resolve SceneAssetHolder dependencies
